Moved DailyFeedingFormsChildHealthAggregationHelper to own file
